pyproject.toml
[project]
name = "caching-sd3"
version = "0.1.0"
description = "Stable Diffusion 3 with latent caching for Modal"
requires-python = ">=3.11"
dependencies = [
    "modal>=0.64.0",
    "torch>=2.1.0",
    "diffusers>=0.30.0",
    "transformers>=4.45.0",
    "accelerate>=0.34.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "matplotlib>=3.7.0",
    "Pillow>=10.0.0",
    "pyyaml>=6.0",
    "pandas>=2.3.3",
    "seaborn>=0.13.2",
]
